LayoutTests/media/modern-media-controls/seek-forward-support/seek-forward-support-expected.txt

Testing the fast forward button in fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media entered fullscreen
PASS media.currentTime is 0
PASS media.paused is true

Media was fast-forwarded for 1 second
PASS media.currentTime is >= 1
PASS media.paused is true

PASS successfullyParsed is true

TEST COMPLETE
